updated changes and added dowloadable test
